requirements.txt
# python==3.11.10
django==5.2
djangorestframework==3.16.1